Added Order

+++ src/de/michaprogs/crm/offer/InsertOffer.java
@@ -68,7 +68,7 @@ public class InsertOffer {
 				count++;
 				ps.setInt(count, mo.getObsListArticle().get(index).getArticleID());
 				count++;
-				ps.setDouble(count, mo.getObsListArticle().get(index).getAmount());
+				ps.setBigDecimal(count, mo.getObsListArticle().get(index).getAmount());
 				count++;
 				ps.setBigDecimal(count, mo.getObsListArticle().get(index).getEk());
 				count++;
